Fix dependency fetch and update
The portion of `sw/make-deps.sh` that tries to update the pslse and
libcxl dependencies is currently broken. Adjust the script to properly
fetch and checkout, or fast-forward pull new changes. Also add a `cd`
line at the top, allowing the script to be called from an arbitrary
working directory.